3.2 Fix adapter
