+++ src/main/java/edu/kis/powp/jobs2d/drivers/adapter/Jobs2dDriverToDrawerAdapter.java
@@ -1,0 +1,42 @@
+package edu.kis.powp.jobs2d.drivers.adapter;
+
+import edu.kis.legacy.drawer.panel.DrawPanelController;
+import edu.kis.legacy.drawer.shape.ILine;
+import edu.kis.legacy.drawer.shape.LineFactory;
+import edu.kis.powp.jobs2d.Job2dDriver;
+import edu.kis.powp.jobs2d.features.DrawerFeature;
+
+/**
+ * driver adapter to drawer with several bugs.
+ */
+public class Jobs2dDriverToDrawerAdapter implements Job2dDriver {
+    private int startX = 0, startY = 0;
+
+	private DrawPanelController drawPanelController;
+
+    public Jobs2dDriverToDrawerAdapter(DrawPanelController drawPanelController) {
+        super();
+        this.drawPanelController = drawPanelController;
+    }
+
+    @Override
+    public void setPosition(int x, int y) {
+        this.startX = x;
+        this.startY = y;
+    }
+
+    @Override
+    public void operateTo(int x, int y) {
+        ILine line = LineFactory.getBasicLine();
+        line.setStartCoordinates(this.startX, this.startY);
+        line.setEndCoordinates(x, y);
+
+        this.drawPanelController.drawLine(line);
+        setPosition(x, y);
+    }
+
+    @Override
+    public String toString() {
+        return "Jobs2dDriverToDrawerAdapter";
+    }
+}
